illuminants_da_técnica_V2_nocontainer/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
Django==1.11.4
djangorestframework==3.6.4
drf-generators==0.3.0
gunicorn==19.7.1
idna==2.6
numpy==1.14.3
olefile==0.45.1
opencv-python==3.4.0.12
Pillow==4.2.1
pytz==2018.4
requests==2.18.4
scikit-learn==0.17.1
scipy==1.1.0
six==1.11.0
urllib3==1.22

